Refactor to store functionality out of card service
