+++ packages/host/app/services/card-service.ts
@@ -1,30 +1,17 @@
 import type Owner from '@ember/owner';
 import Service, { service } from '@ember/service';
 
-import { buildWaiter } from '@ember/test-waiters';
-
-import { isTesting } from '@embroider/macros';
-
-import { stringify } from 'qs';
-
 import { v4 as uuidv4 } from 'uuid';
 
 import {
   SupportedMimeType,
   type LooseCardResource,
-  isSingleCardDocument,
-  isCardCollectionDocument,
   type CardDocument,
   type SingleCardDocument,
   type LooseSingleCardDocument,
   type RealmInfo,
   type Loader,
-  type PatchData,
-  type Relationship,
 } from '@cardstack/runtime-common';
-import type { Query } from '@cardstack/runtime-common/query';
-
-import ENV from '@cardstack/host/config/environment';
 
 import type {
   BaseDef,
@@ -47,31 +34,12 @@ export type CardSaveSubscriber = (
   content: SingleCardDocument | string,
 ) => void;
 
-const { environment } = ENV;
-
-const waiter = buildWaiter('card-service:waiter');
-
 export default class CardService extends Service {
   @service declare private loaderService: LoaderService;
   @service declare private messageService: MessageService;
   @service declare private network: NetworkService;
   @service declare private realm: Realm;
   @service declare private reset: ResetService;
-
-  private async withTestWaiters<T>(cb: () => Promise<T>) {
-    let token = waiter.beginAsync();
-    try {
-      let result = await cb();
-      // only do this in test env--this makes sure that we also wait for any
-      // interior card instance async as part of our ember-test-waiters
-      if (isTesting()) {
-        await this.cardsSettled();
-      }
-      return result;
-    } finally {
-      waiter.endAsync(token);
-    }
-  }
 
   private subscriber: CardSaveSubscriber | undefined;
   // For tracking requests during the duration of this service. Used for being able to tell when to ignore an incremental indexing SSE event.
@@ -106,11 +74,11 @@ export default class CardService extends Service {
     this.loaderToCardAPILoadingCache = new WeakMap();
   }
 
-  onSave(subscriber: CardSaveSubscriber) {
+  _onSave(subscriber: CardSaveSubscriber) {
     this.subscriber = subscriber;
   }
 
-  unregisterSaveSubscriber() {
+  _unregisterSaveSubscriber() {
     this.subscriber = undefined;
   }
 
@@ -209,90 +177,6 @@ export default class CardService extends Service {
     return serialized;
   }
 
-  // TODO move into store
-  async reloadCard(card: CardDef): Promise<void> {
-    // we don't await this in the realm subscription callback, so this test
-    // waiter should catch otherwise leaky async in the tests
-    await this.withTestWaiters(async () => {
-      let incomingDoc: SingleCardDocument = (await this.fetchJSON(
-        card.id,
-        undefined,
-      )) as SingleCardDocument;
-
-      if (!isSingleCardDocument(incomingDoc)) {
-        throw new Error(
-          `bug: server returned a non card document for ${card.id}:
-        ${JSON.stringify(incomingDoc, null, 2)}`,
-        );
-      }
-      let api = await this.getAPI();
-      await api.updateFromSerialized<typeof CardDef>(card, incomingDoc);
-    });
-  }
-
-  // TODO move into store
-  // This is a low-level API that only the store should be calling.
-  // we return undefined if the card changed locally while the save was in-flight.
-  async saveModel(card: CardDef, defaultRealmHref?: string): Promise<void> {
-    let cardChanged = false;
-    function onCardChange() {
-      cardChanged = true;
-    }
-    let token = waiter.beginAsync();
-    let api: typeof CardAPI | undefined;
-    try {
-      api = await this.getAPI();
-      api.subscribeToChanges(card, onCardChange);
-      let doc = await this.serializeCard(card, {
-        // for a brand new card that has no id yet, we don't know what we are
-        // relativeTo because its up to the realm server to assign us an ID, so
-        // URL's should be absolute
-        useAbsoluteURL: true,
-      });
-
-      // send doc over the wire with absolute URL's. The realm server will convert
-      // to relative URL's as it serializes the cards
-      let realmURL = await this.getRealmURL(card);
-      // in the case where we get no realm URL from the card, we are dealing with
-      // a new card instance that does not have a realm URL yet.
-      if (!realmURL) {
-        defaultRealmHref =
-          defaultRealmHref ?? this.realm.defaultWritableRealm?.path;
-        if (!defaultRealmHref) {
-          throw new Error('Could not find a writable realm');
-        }
-        realmURL = new URL(defaultRealmHref);
-      }
-      let json = await this.saveCardDocument(doc, realmURL);
-      let isNew = !card.id;
-
-      // if the card changed while the save was in flight then don't load the
-      // server's version of the card--the next auto save will include these
-      // unsaved changes.
-      if (!cardChanged) {
-        // in order to preserve object equality with the unsaved card instance we
-        // should always use updateFromSerialized()--this way a newly created
-        // instance that does not yet have an id is still the same instance after an
-        // ID has been assigned by the server.
-        await api.updateFromSerialized(card, json);
-      } else if (isNew) {
-        // in this case a new card was created, but there is an immediate change
-        // that was made--so we save off the new ID for the card so in the next
-        // save we'll correlate to the correct card ID
-        card.id = json.data.id;
-      }
-      if (this.subscriber) {
-        this.subscriber(new URL(json.data.id), json);
-      }
-    } catch (err) {
-      console.error(`Failed to save ${card.id}: `, err);
-      throw err;
-    } finally {
-      api?.unsubscribeFromChanges(card, onCardChange);
-      waiter.endAsync(token);
-    }
-  }
-
   async getSource(url: URL) {
     let response = await this.network.authedFetch(url, {
       headers: {
@@ -351,175 +235,6 @@ export default class CardService extends Service {
       throw new Error(errorMessage);
     }
     return response;
-  }
-
-  async patchCard(
-    card: CardDef,
-    doc: LooseSingleCardDocument,
-    patchData: PatchData,
-  ): Promise<void> {
-    let api = await this.getAPI();
-    let linkedCards = await this.loadPatchedCards(patchData, new URL(card.id));
-    for (let [field, value] of Object.entries(linkedCards)) {
-      if (field.includes('.')) {
-        let parts = field.split('.');
-        let leaf = parts.pop();
-        if (!leaf) {
-          throw new Error(`bug: error in field name "${field}"`);
-        }
-        let inner = card;
-        for (let part of parts) {
-          inner = (inner as any)[part];
-        }
-        (inner as any)[leaf.match(/^\d+$/) ? Number(leaf) : leaf] = value;
-      } else {
-        // TODO this could trigger a save. perhaps instead we could
-        // introduce a new option to updateFromSerialized to accept a list of
-        // fields to pre-load? which in this case would be any relationships that
-        // were patched in
-        (card as any)[field] = value;
-      }
-    }
-    await api.updateFromSerialized<typeof CardDef>(card, doc);
-    await this.saveModel(card);
-  }
-
-  private async loadRelationshipCard(rel: Relationship, relativeTo: URL) {
-    if (!rel.links.self) {
-      return;
-    }
-    let id = rel.links.self;
-    let card = await this.getCard(new URL(id, relativeTo).href);
-    return card;
-  }
-
-  // Warning! this is a low level API for getting a card that bypasses the
-  // store's identity map. Cards from here are divorced from the store.
-  async getCard<T extends CardDef = CardDef>(url: URL | string): Promise<T> {
-    if (typeof url === 'string') {
-      url = new URL(url);
-    }
-    let json = await this.fetchJSON(url);
-    if (!isSingleCardDocument(json)) {
-      throw new Error(
-        `bug: server returned a non card document for ${url}:
-      ${JSON.stringify(json, null, 2)}`,
-      );
-    }
-    let card = await this.createFromSerialized(
-      json.data,
-      json,
-      new URL(json.data.id),
-    );
-    return card as T;
-  }
-
-  private async loadPatchedCards(
-    patchData: PatchData,
-    relativeTo: URL,
-  ): Promise<{
-    [fieldName: string]: CardDef | CardDef[];
-  }> {
-    if (!patchData?.relationships) {
-      return {};
-    }
-    let result: { [fieldName: string]: CardDef | CardDef[] } = {};
-    await Promise.all(
-      Object.entries(patchData.relationships).map(async ([fieldName, rel]) => {
-        if (Array.isArray(rel)) {
-          let cards: CardDef[] = [];
-          await Promise.all(
-            rel.map(async (r) => {
-              let card = await this.loadRelationshipCard(r, relativeTo);
-              if (card) {
-                cards.push(card);
-              }
-            }),
-          );
-          result[fieldName] = cards;
-        } else {
-          let card = await this.loadRelationshipCard(rel, relativeTo);
-          if (card) {
-            result[fieldName] = card;
-          }
-        }
-      }),
-    );
-    return result;
-  }
-
-  private async saveCardDocument(
-    doc: LooseSingleCardDocument,
-    realmUrl: URL,
-  ): Promise<SingleCardDocument> {
-    let isSaved = !!doc.data.id;
-    let json = await this.fetchJSON(doc.data.id ?? realmUrl, {
-      method: isSaved ? 'PATCH' : 'POST',
-      body: JSON.stringify(doc, null, 2),
-    });
-    if (!isSingleCardDocument(json)) {
-      throw new Error(
-        `bug: arg is not a card document:
-        ${JSON.stringify(json, null, 2)}`,
-      );
-    }
-    return json;
-  }
-
-  async deleteCard(cardId: string): Promise<void> {
-    if (!cardId) {
-      // the card isn't actually saved yet, so do nothing
-      return;
-    }
-    await this.fetchJSON(cardId, { method: 'DELETE' });
-  }
-
-  // TODO consider retiring this.  i don't think it really does what we want
-  // since it is not live, and the cards that it returns are not live and does
-  // not leverage the identity map from CardResource, so it may create
-  // duplicative instances of cards when it deserializes the results. instead of
-  // using this please use the SearchResource.
-  async search(query: Query, realmURL: URL): Promise<CardDef[]> {
-    let json = await this.fetchJSON(`${realmURL}_search?${stringify(query)}`);
-    if (!isCardCollectionDocument(json)) {
-      throw new Error(
-        `The realm search response was not a card collection document:
-        ${JSON.stringify(json, null, 2)}`,
-      );
-    }
-    let collectionDoc = json;
-    try {
-      console.time('search deserialization');
-      return (
-        await Promise.all(
-          collectionDoc.data.map(async (doc) => {
-            try {
-              return await this.createFromSerialized(
-                doc,
-                collectionDoc,
-                new URL(doc.id),
-              );
-            } catch (e) {
-              console.warn(
-                `Skipping ${
-                  doc.id
-                }. Encountered error deserializing from search result for query ${JSON.stringify(
-                  query,
-                  null,
-                  2,
-                )} against realm ${realmURL}`,
-                e,
-              );
-              return undefined;
-            }
-          }),
-        )
-      ).filter(Boolean) as CardDef[];
-    } finally {
-      if (environment !== 'test') {
-        console.timeEnd('search deserialization');
-      }
-    }
   }
 
   async getFields(
