+++ packages/host/app/services/operator-mode-state-service.ts
@@ -181,7 +181,7 @@ export default class OperatorModeStateService extends Service {
           document.data.relationships = mergedRel;
         }
       }
-      await this.cardService.patchCard(card, document, patch);
+      await this.store.patchCard(card, document, patch);
     }
   });
 
@@ -193,7 +193,7 @@ export default class OperatorModeStateService extends Service {
       throw new Error(`Could not determine the realm for card "${cardId}"`);
     }
 
-    await this.cardService.deleteCard(cardId);
+    await this.store.deleteCard(cardId);
 
     // remove all stack items for the deleted card
     let items: StackItem[] = [];
